+++ compiler-core/src/erlang/tests/bit_arrays.rs
@@ -526,3 +526,28 @@ pub fn go(wibble: String) {
 "#
     )
 }
+
+// https://github.com/gleam-lang/gleam/issues/5208
+#[test]
+fn unit_option_ignores_bytes() {
+    assert_erl!(
+        "
+pub fn main() {
+  let assert <<x:unit(2)-bytes-size(3)>> = <<1:6>>
+  x
+}
+"
+    );
+}
+
+#[test]
+fn unit_option_ignores_bytes_regardless_of_order() {
+    assert_erl!(
+        "
+pub fn main() {
+  let assert <<x:bytes-unit(2)-size(3)>> = <<1:6>>
+  x
+}
+"
+    );
+}